renamed structure_based module to ligand_receptor
