+++ openpharmacophore/__init__.py
@@ -11,6 +11,7 @@ from ._pyunitwizard import puw
 from .pharmacophore import PharmacophoricPoint, distance_between_pharmacophoric_points
 from .pharmacophore import LigandBasedPharmacophore
 from .pharmacophore import LigandReceptorPharmacophore
+from .pharmacophore import load_from_file
 from .screening import VirtualScreening
