+++ openpharmacophore/pharmacophore/__init__.py
@@ -1,3 +1,4 @@
 from .pharmacophoric_point import PharmacophoricPoint, distance_between_pharmacophoric_points
 from .ligand_based import LigandBasedPharmacophore
-from .structure_based import LigandReceptorPharmacophore
+from .ligand_receptor import LigandReceptorPharmacophore
+from .load_from_file import load_from_file
